+++ src/result.js
@@ -3,17 +3,23 @@
 const SimpleDOM = require('simple-dom');
 const HTMLSerializer = new SimpleDOM.HTMLSerializer(SimpleDOM.voidMap);
 
+const SHOEBOX_TAG_PATTERN = '<script type="fastboot/shoebox"';
+const HTML_HEAD_REGEX = /^([\s\S]*<\/head>)([\s\S]*)/;
+
 /**
  * Represents the rendered result of visiting an Ember app at a particular URL.
  * A `Result` object is returned from calling {@link FastBoot}'s `visit()`
  * method.
  */
 class Result {
-  constructor(options) {
-    this._instanceDestroyed = false;
-    this._doc = options.doc;
-    this._html = options.html;
-    this._fastbootInfo = options.fastbootInfo;
+  constructor(doc, html, fastbootInfo) {
+    this.isDestroyed = false;
+
+    this._doc = doc;
+    this._html = html;
+    this._fastbootInfo = fastbootInfo;
+    this.applicationInstance = undefined;
+    this.applicationInstanceInstance = undefined;
   }
 
   /**
@@ -42,7 +48,55 @@ class Result {
       }
     }
 
-    return insertIntoIndexHTML(this._html, this._head, this._body);
+    return insertIntoIndexHTML(
+      this._html,
+      this._htmlAttributes,
+      this._head,
+      this._body,
+      this._bodyAttributes
+    );
+  }
+
+  /**
+   * Returns the HTML representation of the rendered route, inserted
+   * into the application's `index.html`, split into chunks.
+   * The first chunk contains the document's head, the second contains the body
+   * until just before the shoebox tags (if there are any) and the last chunk
+   * contains the shoebox tags and the closing `body` tag. If there are no
+   * shoebox tags, there are only 2 chunks and the second one contains the
+   * complete document body, including the closing `body` tag.
+   *
+   * @returns {Promise<Array<String>>} the application's DOM serialized to HTML, split into chunks
+   */
+  chunks() {
+    return insertIntoIndexHTML(
+      this._html,
+      this._htmlAttributes,
+      this._head,
+      this._body,
+      this._bodyAttributes
+    ).then(html => {
+      let docParts = html.match(HTML_HEAD_REGEX);
+      if (!docParts || docParts.length === 1) {
+        return [html];
+      }
+
+      let [, head, body] = docParts;
+
+      if (!head || !body) {
+        throw new Error(
+          'Could not idenfity head and body of the document! Make sure the document is well formed.'
+        );
+      }
+
+      let [plainBody, ...shoeboxes] = body.split(SHOEBOX_TAG_PATTERN);
+
+      let chunks = [head, plainBody].concat(
+        shoeboxes.map(shoebox => `${SHOEBOX_TAG_PATTERN}${shoebox}`)
+      );
+
+      return chunks;
+    });
   }
 
   /**
@@ -53,28 +107,28 @@ class Result {
   domContents() {
     return {
       head: this._head,
-      body: this._body
+      body: this._body,
     };
   }
 
   /**
    * @private
    *
-   * Called once the Result has finished being constructed and the application
-   * instance has finished rendering. Once `finalize()` is called, state is
-   * gathered from the completed application instance and statically copied
-   * to this Result instance.
+   * Called once the Result has finished being constructed and the
+   * ApplicationInstance instance has finished rendering. Once `finalize()` is
+   * called, state is gathered from the completed ApplicationInstance instance
+   * and statically copied to this Result instance.
    */
   _finalize() {
     if (this.finalized) {
-      throw new Error("Results cannot be finalized more than once");
+      throw new Error('Results cannot be finalized more than once');
     }
 
     // Grab some metadata from the sandboxed application instance
     // and copy it to this Result object.
-    let instance = this.instance;
-    if (instance) {
-      this._finalizeMetadata(instance);
+    let { applicationInstanceInstance } = this;
+    if (applicationInstanceInstance) {
+      this._finalizeMetadata(applicationInstanceInstance);
     }
 
     this._finalizeHTML();
@@ -83,12 +137,12 @@ class Result {
     return this;
   }
 
-  _finalizeMetadata(instance) {
-    if (instance._booted) {
-      this.url = instance.getURL();
+  _finalizeMetadata(applicationInstanceInstance) {
+    if (applicationInstanceInstance._booted) {
+      this.url = applicationInstanceInstance.getURL();
     }
 
-    let response = this._fastbootInfo.response;
+    let { response } = this._fastbootInfo;
 
     if (response) {
       this.headers = response.headers;
@@ -96,18 +150,40 @@ class Result {
     }
   }
 
-  _destroyAppInstance() {
-    if (this.instance && !this._instanceDestroyed) {
-      this._instanceDestroyed = true;
-      this.instance.destroy();
-      return true;
+  _destroy() {
+    if (this.isDestroyed) {
+      return false;
     }
-    return false;
+
+    this.isDestroyed = true;
+
+    if (this.applicationInstanceInstance !== undefined) {
+      this.applicationInstanceInstance.destroy();
+    }
+
+    if (this.applicationInstance !== undefined) {
+      this.applicationInstance.destroy();
+    }
+
+    return true;
   }
 
   _finalizeHTML() {
+    let htmlElement = this._doc.documentElement;
     let head = this._doc.head;
     let body = this._doc.body;
+
+    if (htmlElement.attributes.length > 0) {
+      this._htmlAttributes = HTMLSerializer.attributes(htmlElement.attributes);
+    } else {
+      this._htmlAttributes = null;
+    }
+
+    if (body.attributes.length > 0) {
+      this._bodyAttributes = HTMLSerializer.attributes(body.attributes);
+    } else {
+      this._bodyAttributes = null;
+    }
 
     if (head) {
       head = HTMLSerializer.serializeChildren(head);
@@ -116,20 +192,26 @@ class Result {
     body = HTMLSerializer.serializeChildren(body);
 
     this._head = head;
-    this._body = body;
+
+    // Adding script boundary around the body
+    this._body = `<script type="x/boundary" id="fastboot-body-start"></script>${body}<script type="x/boundary" id="fastboot-body-end"></script>`;
   }
 }
 
 function missingTag(tag) {
-  return Promise.reject(new Error(`Fastboot was not able to find ${tag} in base HTML. It could not replace the contents.`));
+  throw new Error(
+    `Fastboot was not able to find ${tag} in base HTML. It could not replace the contents.`
+  );
 }
 
-function insertIntoIndexHTML(html, head, body) {
-  if (!html) { return Promise.resolve(html); }
+async function insertIntoIndexHTML(html, htmlAttributes, head, body, bodyAttributes) {
+  if (!html) {
+    return Promise.resolve(html);
+  }
   let isBodyReplaced = false;
   let isHeadReplaced = false;
 
-  html = html.replace(/<\!-- EMBER_CLI_FASTBOOT_(HEAD|BODY) -->/g, function(match, tag) {
+  html = html.replace(/<!-- EMBER_CLI_FASTBOOT_(HEAD|BODY) -->/g, function(match, tag) {
     if (tag === 'HEAD' && head && !isHeadReplaced) {
       isHeadReplaced = true;
       return head;
@@ -140,14 +222,26 @@ function insertIntoIndexHTML(html, head, body) {
     return '';
   });
 
-  if (head && !isHeadReplaced) {
-    return missingTag('<!--EMBER_CLI_FASTBOOT_HEAD-->');
-  }
-  if (body && !isBodyReplaced) {
-    return missingTag('<!--EMBER_CLI_FASTBOOT_BODY-->');
+  if (htmlAttributes) {
+    html = html.replace(/<html[^>]*/i, function(match) {
+      return match + ' ' + htmlAttributes;
+    });
   }
 
-  return Promise.resolve(html);
+  if (bodyAttributes) {
+    html = html.replace(/<body[^>]*/i, function(match) {
+      return match + ' ' + bodyAttributes;
+    });
+  }
+
+  if (head && !isHeadReplaced) {
+    missingTag('<!--EMBER_CLI_FASTBOOT_HEAD-->');
+  }
+  if (body && !isBodyReplaced) {
+    missingTag('<!--EMBER_CLI_FASTBOOT_BODY-->');
+  }
+
+  return html;
 }
 
 module.exports = Result;